Fix & Function Implementation
Fixes -> When searching for GameObjects using the Unity::GameObject::Find function, it will first search for the exact thing you gave the argument; if it fails, it will do another search but will add "(Clone)" to the end of the name.

Implementation(s)

Added function called "SetTag", it calls a function that belongs to the GameObject class that sets the GameObjects tag as whatever you entered into the arguments.