Implemented new functions
Unity Functions added:

Set Name -> Able to set the name of a gameobject.

Get Tag -> Able to get the tag of a gameobject.

Set Tag -> Able to set tag of a gameobject.

DontDestroyOnLoad -> Able to call DontDestroyOnLoad function on gameobjects, makes sure it doesnt get destroyed on loading new scenes.

Get Freeze Rotation -> Able to get the boolean of Freeze Rotation on Rigidbodies.

Set Freeze Rotation -> Able to set the boolean of Freeze Rotation on Rigidbodies.

Set Is Kinematic -> Able to set the Is Kinematic boolean on Rigidbodies.

Get Is Kinematic -> Able to get the Is Kinematic boolean on Rigidbodies.